Making the code public.
Until today this project has existed only on my laptop.  Since I'm
changing the subject of my dissertation to one that overlaps with what
I do in my full-time job, I won't be maintaining or developing this,
but I'd like it to be available for others to use or extend.